Update last-modified-timestamp.php
Adds the ability to pass a post or page ID as a parameter and display its last modified timestamp instead. For example:

`[last-modified post="7"]`

If no parameter is passed, or the parameter is invalid (i.e. the post doesn't exist), the shortcode reverts to its default behavior of displaying the last modified timestamp for the post it's embedded in.

This PR also bumps the plugin version from 1.0.5 to 1.1.

Closes #11.